voice_agent/requirements.txt
livekit-agents
livekit-plugins-deepgram
livekit-plugins-cartesia
livekit-plugins-openai
livekit-plugins-silero
python-dotenv
redis
fastapi
uvicorn
pyyaml
aiohttp
twilio

